much better.
needs some defuzz
